bpa_lab_warehouse_operations_docker_version/TXT_Controller_Apps/requirements.txt
paho-mqtt
ftrobopy